+++ usr/sys/h/fblk.h
@@ -1,5 +1,5 @@
 struct fblk
 {
-	int    	df_nfree;
+	i16    	df_nfree;
 	daddr_t	df_free[NICFREE];
 };
